Change images based on weather and time
The app now changes it's images when the weather and time changes, we
will use different set of images in the short future, but these will
work for now.
Also added a snackbar to show when the mqtt connection options are not
correct or complete, to make it easier for the user to change them.